Added a disclaimer for several learning objectives.
Some of the learning objectives are not 100% in keeping with the course design as it evolved during the summer of 2012. Instead of inserting the objectives developed during the summer, the template/boilerplate departmental objectives were used.

However, several are hit less directly in the course. To indicate that, a small disclaimer was added.