Updated syllabus.
This reflects a more student-centered approach, and treats the syllabus as more of a "learning object." That is, it is less of a "contract" document, and is an attempt to be more of a "teaching" document.

Kudos to Jan for sharing her syllabus (from which portions are ripped/modified), and helping me see how it doesn't have to be an advesarial/"contractual" framing of course policies.